pyproject.toml
[tool.poetry]
name = "popugprojects"
version = "0.1.0"
description = ""
authors = ["Bpem9 <itimeq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
dj-database-url = "^2.0.0"
django-oauth-toolkit = "^2.3.0"
django-cors-headers = "^4.2.0"
psycopg2-binary = "^2.9.7"
django-rest-registration = "^0.8.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
